# Faster UTF-8 decoding from mixed buffers