Part of https://github.com/scala-native/scala-native/issues/4280

UTF-8 decoding had two implementations, a naive `decodeLoopNoArray` and an optimized `decodeLoopArray` (for when both Buffers are backed by an array).

In real world applications it's to use the buffers as a bridge between the "Native" and "Scala" worlds, so one of the buffers is backed by pointers, while another is backed by arrays (e.g. in `fromCString`). This case was not being optimized.

This PR merges both implementations into a single one, so that `Ptr->Array` and `Array->Ptr` conversions are also optimized. This also opens the door for future optimizations using the extended`Buffer#hasPointer(): Boolean` / `Buffer#pointer(): unsafe.Ptr[Byte]` operations.

While this adds a few extra ifs in the `ArrayBuffer->ArrayBuffer` conversion, I didn't see any meaningful performance degradation, so I just removed the specialized method, but it would be trivial to add it back again.

I used the following naive code for benchmarks:

```scala
package example

import java.nio.*
import java.nio.charset.Charset
import scala.util.Random
import scala.scalanative.memory.PointerBuffer
import scalanative.unsafe.*

object Main {
  val charset = Charset.forName("UTF-8")
  val encoder = charset.newEncoder()
  val decoder = charset.newDecoder()

  def time[T](task: String)(f: => T): T =
    println(s"Starting $task")
    val start = System.currentTimeMillis()
    val res = f
    println(s"Done in ${System.currentTimeMillis() - start}ms")
    res

  @main def run() = {
    val data: Array[Byte] =
      time("fill array")(("TEST: 👍 | " + Random.alphanumeric.take(20 * 1024 * 1024).mkString).getBytes(charset))

    val ptr = data.at(0)

    val arrRes = time("decode array"){
      val inputBuffer = ByteBuffer.wrap(data, 0, data.size)
      decoder.decode(inputBuffer)
    }

    val ptrRes = time("decode ptr"){
      val inputBuffer = PointerBuffer.wrap(ptr, data.size)
      decoder.decode(inputBuffer)
    }

    // I had some issues running the unit tests locally, so I added some sanity checks here.
    // Hopefully, the CI will be green.
    println(arrRes.array().mkString("") == ptrRes.array().mkString(""))
    println(arrRes.array().mkString("").take(16))
  }
}
```

On 0.5.7:
```
Starting fill array
Done in 12127ms
Starting decode array
Done in 25ms
Starting decode ptr
Done in 88ms
true
TEST: 👍 | 5HIw5
```

With this PR:
```
Starting fill array
Done in 12092ms
Starting decode array
Done in 27ms
Starting decode ptr
Done in 36ms
true
TEST: 👍 | dHrUG
```