@LeeTibbert 2025-04-06
JD557 Looks like you have advanced the peg and removed a large chunk of code in the process.

Curious, any guesses where the 4%, 2 millisecond, decrease in speed of the "decode array" path 
comes from?  2 milliseconds over the course of 20 MB is not a huge difference. Seems like
more that a few 'if' tests would cause, even over thousands of iterations.  

If I understand you correctly the "decode array" path is much less used over a large mix of
workloads, than the "decode pointer" path, so there is an overall win, subject to the
usual disclaimer for any change "but my workload regressed".

I looked at the code "at arms length" and saw the code reduction. I have to wait for the day
to cool off before I can study it in detail.  

The change looks good and I do not want to get in the way of it being merged whilst I 
puzzle about milliseconds.   

The reason for my at least reading level interest is that  I hope to bring some  Java 9(?) `mismatch` 
implementations to `Buffers` sometime before the Millennium and the speed of the underlying
`Arrays.mismatch` implementations will be important, even if only measured in milliseconds.