@JD557 2025-04-06
> Curious, any guesses where the 4%, 2 millisecond, decrease in speed of the "decode array" path
comes from? 2 milliseconds over the course of 20 MB is not a huge difference. Seems like
more that a few 'if' tests would cause, even over thousands of iterations.

This test was not very scientific, and the numbers varied a bit between tests, so I'm assuming it was just that

> If I understand you correctly the "decode array" path is much less used over a large mix of
workloads, than the "decode pointer" path, so there is an overall win, subject to the
usual disclaimer for any change "but my workload regressed".

To be fair, I think this might not be true *yet*. It would be true after some changes that I would like to do to `fromCString`.