@LeeTibbert 2025-04-06
>  changes that I would like to do to fromCString

Yeah!  Looking forward to seeing them when, in the fullness of time, they arrive.

Improvements there are likely to have benefits all across SN.