@LeeTibbert 2025-04-06
Oh, could I please use, with attribution,  bits or variants of the code you provided in the base Issue
to create unit-tests?   That would expedite implementation.