@JD557 2025-04-06
> Oh, could I please use bits or variants of the code you provided in the base Issue?

Of course!